algorithms/java/src/test/java/org/jessenpan/leetcode/array/S999AvailableCapturesForRookTest.java
package org.jessenpan.leetcode.array;

import org.junit.Test;

import static org.junit.Assert.assertEquals;

/**
 * @author jessenpan
 */
public class S999AvailableCapturesForRookTest {

    private S999AvailableCapturesForRook availableCapturesForRook = new S999AvailableCapturesForRook();

    @Test
    public void test() {
        int numRook = availableCapturesForRook.numRookCaptures(new char[][] { { '.', '.', '.', '.', '.', '.', '.', '.' }, { '.', '.', '.', 'p', '.', '.', '.', '.' }, { '.', '.', '.', 'R', '.', '.', '.', 'p' }, { '.', '.', '.', '.', '.', '.', '.', '.' }, { '.', '.', '.', '.', '.', '.', '.', '.' }, { '.', '.', '.', 'p', '.', '.', '.', '.' }, { '.', '.', '.', '.', '.', '.', '.', '.' }, { '.', '.', '.', '.', '.', '.', '.', '.' } });
        assertEquals(3,numRook);
    }

    @Test
    public void test1() {
        int numRook = availableCapturesForRook.numRookCaptures(new char[][] {{'.','.','.','.','.','.','.','.'},{'.','p','p','p','p','p','.','.'},{'.','p','p','B','p','p','.','.'},{'.','p','B','R','B','p','.','.'},{'.','p','p','B','p','p','.','.'},{'.','p','p','p','p','p','.','.'},{'.','.','.','.','.','.','.','.'},{'.','.','.','.','.','.','.','.'}});
        assertEquals(0,numRook);
    }

}
